# docs/swagger.yaml
definitions:
  dto.TaskReqDto:
    properties:
      name:
        type: string
      status:
        maximum: 1
        minimum: 0
        type: integer
    required:
    - name
    - status
    type: object
  http_response.ResponseBody:
    properties:
      data:
        x-omitempty: true
      message:
        type: string
      status:
        type: string
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /task/:id:
    get:
      consumes:
      - application/json
      description: get task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_response.ResponseBody'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_response.ResponseBody'
      summary: get task
      tags:
      - Task
  /tasks:
    get:
      consumes:
      - application/json
      description: get all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_response.ResponseBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Task'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_response.ResponseBody'
      summary: get all tasks
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: create new task
      parameters:
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TaskReqDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_response.ResponseBody'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_response.ResponseBody'
      summary: Create new task
      tags:
      - Task
  /tasks/:id:
    delete:
      consumes:
      - application/json
      description: deletee task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_response.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_response.ResponseBody'
      summary: delete task
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: update task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: body params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TaskReqDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_response.ResponseBody'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_response.ResponseBody'
      summary: update task
      tags:
      - Task
swagger: "2.0"